+++ src/test/java/org/opensearch/flowframework/workflow/CreateSearchPipelineStepTests.java
@@ -11,12 +11,10 @@ package org.opensearch.flowframework.workflow;
 import org.opensearch.action.search.PutSearchPipelineRequest;
 import org.opensearch.action.support.PlainActionFuture;
 import org.opensearch.action.support.master.AcknowledgedResponse;
-import org.opensearch.action.update.UpdateResponse;
 import org.opensearch.client.AdminClient;
 import org.opensearch.client.Client;
 import org.opensearch.client.ClusterAdminClient;
 import org.opensearch.core.action.ActionListener;
-import org.opensearch.core.index.shard.ShardId;
 import org.opensearch.flowframework.indices.FlowFrameworkIndicesHandler;
 import org.opensearch.test.OpenSearchTestCase;
 
@@ -27,9 +25,7 @@ import java.util.concurrent.ExecutionException;
 
 import org.mockito.ArgumentCaptor;
 
-import static org.opensearch.action.DocWriteResponse.Result.UPDATED;
 import static org.opensearch.flowframework.common.CommonValue.CONFIGURATIONS;
-import static org.opensearch.flowframework.common.CommonValue.WORKFLOW_STATE_INDEX;
 import static org.opensearch.flowframework.common.WorkflowResources.MODEL_ID;
 import static org.opensearch.flowframework.common.WorkflowResources.PIPELINE_ID;
 import static org.mockito.ArgumentMatchers.any;
@@ -43,7 +39,7 @@ import static org.mockito.Mockito.when;
 public class CreateSearchPipelineStepTests extends OpenSearchTestCase {
 
     private WorkflowData inputData;
-    private WorkflowData outpuData;
+    private WorkflowData outputData;
     private Client client;
     private AdminClient adminClient;
     private ClusterAdminClient clusterAdminClient;
@@ -63,7 +59,7 @@ public class CreateSearchPipelineStepTests extends OpenSearchTestCase {
         );
 
         // Set output data to returned pipelineId
-        outpuData = new WorkflowData(Map.ofEntries(Map.entry(PIPELINE_ID, "pipelineId")), "test-id", "test-node-id");
+        outputData = new WorkflowData(Map.ofEntries(Map.entry(PIPELINE_ID, "pipelineId")), "test-id", "test-node-id");
 
         client = mock(Client.class);
         adminClient = mock(AdminClient.class);
@@ -78,10 +74,10 @@ public class CreateSearchPipelineStepTests extends OpenSearchTestCase {
         CreateSearchPipelineStep createSearchPipelineStep = new CreateSearchPipelineStep(client, flowFrameworkIndicesHandler);
 
         doAnswer(invocation -> {
-            ActionListener<UpdateResponse> updateResponseListener = invocation.getArgument(4);
-            updateResponseListener.onResponse(new UpdateResponse(new ShardId(WORKFLOW_STATE_INDEX, "", 1), "id", -2, 0, 0, UPDATED));
+            ActionListener<WorkflowData> updateResponseListener = invocation.getArgument(4);
+            updateResponseListener.onResponse(new WorkflowData(Map.of(PIPELINE_ID, "pipelineId"), "test-id", "test-node-id"));
             return null;
-        }).when(flowFrameworkIndicesHandler).updateResourceInStateIndex(anyString(), anyString(), anyString(), anyString(), any());
+        }).when(flowFrameworkIndicesHandler).addResourceToStateIndex(any(WorkflowData.class), anyString(), anyString(), anyString(), any());
 
         @SuppressWarnings("unchecked")
         ArgumentCaptor<ActionListener<AcknowledgedResponse>> actionListenerCaptor = ArgumentCaptor.forClass(ActionListener.class);
@@ -100,7 +96,7 @@ public class CreateSearchPipelineStepTests extends OpenSearchTestCase {
         actionListenerCaptor.getValue().onResponse(new AcknowledgedResponse(true));
 
         assertTrue(future.isDone());
-        assertEquals(outpuData.getContent(), future.get().getContent());
+        assertEquals(outputData.getContent(), future.get().getContent());
     }
 
     public void testCreateSearchPipelineStepFailure() throws InterruptedException {